Word Ladder Game in Java
A Word Ladder is the connection between 2 same letter words, such that
each transition to a step can only be made by replacing only 1 letter in
the source word and the target word is always a meaning word. A good
solution has a number of steps equal to the number of places the given
words differ in (Hamming Distance). E.g.
Cold -> Cord -> Card -> Ward -> Warm
If a good solution is not possible, then look towards the next best
solution, i.e. shortest chain in the word ladder.